Changed TIM2 prescaler to achieve 1MHz timing. Added some pin toggles to code to test timing.
Currently, the loopFOC is called at a freq of 25 kHz, and move is called at 25Hz.

The behavior of the motor is hard to stabilise - probably due to PID values. Also, and considering to what extent encoder interrupt servicing is delaying various loops. Am looking to switch encoder inputs to a dedicated timer setup in "encoder" mode.